+++ packages/framework/src/Command/ChangeEnvironmentCommand.php
@@ -72,6 +72,6 @@ class ChangeEnvironmentCommand extends Command
 
         $output->writeln(sprintf('Application environment successfully changed to "%s".', $targetEnvironment));
 
-        return CommandResultCodes::RESULT_OK;
+        return Command::SUCCESS;
     }
 }