+++ packages/migrations/src/Command/GetCountOfMigrationsToExecuteCommand.php
@@ -7,10 +7,13 @@ use Doctrine\Migrations\Version\AliasResolver;
 use Shopsys\MigrationBundle\Component\Doctrine\Migrations\MigrationLockPlanCalculator;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 class GetCountOfMigrationsToExecuteCommand extends Command
 {
+    protected const OPTION_SIMPLE = 'simple';
+
     /**
      * @var string
      */
@@ -41,7 +44,13 @@ class GetCountOfMigrationsToExecuteCommand extends Command
     protected function configure(): void
     {
         $this
-            ->setDescription('Get count of migrations to execute.');
+            ->setDescription('Get count of migrations to execute.')
+            ->addOption(
+                static::OPTION_SIMPLE,
+                null,
+                InputOption::VALUE_NONE,
+                'Return only count of migrations to execute.',
+            );
     }
 
     /**
@@ -52,8 +61,12 @@ class GetCountOfMigrationsToExecuteCommand extends Command
         $latestVersion = $this->aliasResolver->resolveVersionAlias('latest');
         $migrationPlanList = $this->migrationLockPlanCalculator->getPlanUntilVersion($latestVersion);
 
-        $output->writeln('Count of migrations to execute: ' . $migrationPlanList->count());
+        if ($input->getOption(static::OPTION_SIMPLE)) {
+            $output->writeln((string)$migrationPlanList->count());
+        } else {
+            $output->writeln('Count of migrations to execute: ' . $migrationPlanList->count());
+        }
 
-        return 0;
+        return Command::SUCCESS;
     }
 }